Initial testHashTransaction tests
